requirements.txt
fastapi
uvicorn[standard]
python-multipart
groq
python-dotenv
pdfplumber
python-docx
pdf2image
pytesseract
reportlab